requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
discord.py>=2.0.0
requests>=2.28.0
inifile>=0.5.0